rknn2/pyproject.toml
[project]
name = "rknn2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rknn-toolkit2>=2.3.2",
    "setuptools>=80.9.0",
    "singtown-ai==1.6.2",
]
